Support decimal values for Annual Paid Days (#12)
* Support decimal values for Annual Paid Days (PTO allowance)

Changes int to decimal(18,2) across the model, ViewModel, balance
calculation, and input so consultants can be assigned fractional
PTO days (e.g. 15.5). Also gitignores launchSettings.json and
CLAUDE.md to prevent local dev secrets from being committed.

* Allow any decimal value for Annual Paid Days input

---------

+++ OceansApp/OceansApp.DataAccess/Repository/TimeOffRequestRepository.cs
@@ -42,8 +42,8 @@ namespace OceansApp.DataAccess.Repository
 
             if (consultant.IsEligibleForPaidTimeOff && consultant.AnnualPaidTimeOffDays.HasValue)
             {
-                int annualDays = consultant.AnnualPaidTimeOffDays.Value;
-                int ptoAllowance = annualDays;
+                decimal annualDays = consultant.AnnualPaidTimeOffDays.Value;
+                decimal ptoAllowance = annualDays;
                 int ptoUsedAndPending = await GetUsedDaysAsync(consultantId, "PTO", currentYear);
                 result.PtoAvailable = ptoAllowance - ptoUsedAndPending;
             }
